refactor: refactor packages and create verification options (#129)
